// proto/transfer_v2.ts
export enum Action {
  UNKNOWN = 0,
  /** KEEP - files are already same, keep it */
  KEEP = 1,
  /** DELETE - extra file in your system, delete it */
  DELETE = 2,
  /** COMPLETE - sending files is complete, you can close the write stream */
  COMPLETE = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface FileRequest {
  info: FileInfo[];
}

export interface FileInfo {
  path: string;
  hash: string;
}

export interface FileResponse {
  path: string;
  action?: Action | undefined;
  data?: Uint8Array | undefined;
}

function createBaseFileRequest(): FileRequest {
  return { info: [] };
}

export const FileRequest: MessageFns<FileRequest> = {
  encode(message: FileRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.info) {
      FileInfo.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FileRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFileRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.info.push(FileInfo.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FileRequest {
    return { info: globalThis.Array.isArray(object?.info) ? object.info.map((e: any) => FileInfo.fromJSON(e)) : [] };
  },

  toJSON(message: FileRequest): unknown {
    const obj: any = {};
    if (message.info?.length) {
      obj.info = message.info.map((e) => FileInfo.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FileRequest>, I>>(base?: I): FileRequest {
    return FileRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FileRequest>, I>>(object: I): FileRequest {
    const message = createBaseFileRequest();
    message.info = object.info?.map((e) => FileInfo.fromPartial(e)) || [];
    return message;
  },
};

function createBaseFileInfo(): FileInfo {
  return { path: "", hash: "" };
}

export const FileInfo: MessageFns<FileInfo> = {
  encode(message: FileInfo, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.path !== "") {
      writer.uint32(10).string(message.path);
    }
    if (message.hash !== "") {
      writer.uint32(18).string(message.hash);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FileInfo {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFileInfo();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.path = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.hash = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FileInfo {
    return {
      path: isSet(object.path) ? globalThis.String(object.path) : "",
      hash: isSet(object.hash) ? globalThis.String(object.hash) : "",
    };
  },

  toJSON(message: FileInfo): unknown {
    const obj: any = {};
    if (message.path !== "") {
      obj.path = message.path;
    }
    if (message.hash !== "") {
      obj.hash = message.hash;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FileInfo>, I>>(base?: I): FileInfo {
    return FileInfo.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FileInfo>, I>>(object: I): FileInfo {
    const message = createBaseFileInfo();
    message.path = object.path ?? "";
    message.hash = object.hash ?? "";
    return message;
  },
};

function createBaseFileResponse(): FileResponse {
  return { path: "", action: undefined, data: undefined };
}

export const FileResponse: MessageFns<FileResponse> = {
  encode(message: FileResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.path !== "") {
      writer.uint32(10).string(message.path);
    }
    if (message.action !== undefined) {
      writer.uint32(16).int32(message.action);
    }
    if (message.data !== undefined) {
      writer.uint32(26).bytes(message.data);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FileResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFileResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.path = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.action = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.data = reader.bytes();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FileResponse {
    return {
      path: isSet(object.path) ? globalThis.String(object.path) : "",
      action: isSet(object.action) ? actionFromJSON(object.action) : undefined,
      data: isSet(object.data) ? bytesFromBase64(object.data) : undefined,
    };
  },

  toJSON(message: FileResponse): unknown {
    const obj: any = {};
    if (message.path !== "") {
      obj.path = message.path;
    }
    if (message.action !== undefined) {
      obj.action = actionToJSON(message.action);
    }
    if (message.data !== undefined) {
      obj.data = base64FromBytes(message.data);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<FileResponse>, I>>(base?: I): FileResponse {
    return FileResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<FileResponse>, I>>(object: I): FileResponse {
    const message = createBaseFileResponse();
    message.path = object.path ?? "";
    message.action = object.action ?? undefined;
    message.data = object.data ?? undefined;
    return message;
  },
};

export type FileServiceService = typeof FileServiceService;
export const FileServiceService = {
  syncFiles: {
    path: "/transfer_v2.FileService/SyncFiles",
    requestStream: false,
    responseStream: true,
    requestSerialize: (value: FileRequest) => Buffer.from(FileRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => FileRequest.decode(value),
    responseSerialize: (value: FileResponse) => Buffer.from(FileResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => FileResponse.decode(value),
  },
} as const;

export interface FileServiceServer extends UntypedServiceImplementation {
  syncFiles: handleServerStreamingCall<FileRequest, FileResponse>;
}

export interface FileServiceClient extends Client {
  syncFiles(request: FileRequest, options?: Partial<CallOptions>): ClientReadableStream<FileResponse>;
  syncFiles(
    request: FileRequest,
    metadata?: Metadata,
    options?: Partial<CallOptions>,
  ): ClientReadableStream<FileResponse>;
}

export const FileServiceClient = makeGenericClientConstructor(
  FileServiceService,
  "transfer_v2.FileService",
) as unknown as {
  new (address: string, credentials: ChannelCredentials, options?: Partial<ClientOptions>): FileServiceClient;
  service: typeof FileServiceService;
  serviceName: string;
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}